Fully functioning with js confirmation of the delete link

+++ WebContent/WEB-INF/view/list-customers.jsp
@@ -50,7 +50,7 @@
 						<td> ${tempCustomer.email}</td>
 						<td>
 							<!-- display the update link -->
-							<a href="${updateLink}">Update</a> | <a href="${deleteLink}">Delete</a>
+							<a href="${updateLink}">Update</a> | <a href="${deleteLink}" onclick="if (!(confirm('Are you sure you want to delete this customer?'))) return false;">Delete</a>
 						</td>
 					</tr>
